[python] add yhat, fix cp&ap axis title
