+++ python/dalex/dalex/__init__.py
@@ -1,5 +1,9 @@
+# dx.Explainer
 from ._explainer.object import Explainer
+
+# from dalex.datasets import load_titanic
 from . import datasets
+
 
 __all__ = [
     "Explainer",
